# cli: update progress output after stream has ended